# src/app/api/v1/auth/session/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { createRouteHandlerClient } from '@/lib/supabase-server';

export async function GET(request: NextRequest) {
  try {
    const supabase = createRouteHandlerClient();

    // Get the current user session
    const { data: { session }, error: sessionError } = await supabase.auth.getSession();

    if (sessionError) {
      return NextResponse.json(
        { error: 'Failed to get session' },
        { status: 500 }
      );
    }

    if (!session) {
      return NextResponse.json(
        {
          user: null,
          session: null,
          subscription_status: null
        },
        { status: 200 }
      );
    }

    // Get user profile information
    const { data: profile } = await supabase
      .from('user_profiles')
      .select('*')
      .eq('id', session.user.id)
      .single();

    // Get subscription status
    const { data: subscription } = await supabase
      .from('subscriptions')
      .select('status')
      .eq('user_id', session.user.id)
      .maybeSingle();

    return NextResponse.json({
      user: {
        id: session.user.id,
        email: session.user.email,
        profile: profile || null
      },
      session: {
        access_token: session.access_token,
        expires_at: session.expires_at
      },
      subscription_status: subscription ? (subscription as any).status : null
    });

  } catch (error) {
    console.error('Session API error:', error);
    return NextResponse.json(
      { error: 'Internal server error' },
      { status: 500 }
    );
  }
}